Workaround for kernel panic when using counter groups
We were apparently not setting up counter groups correctly, this
caused a kernel panic (at least on 2.6.36) when running perfrecord and
caused rawperf to return 0 for all counters except for the leader.
The old method when setting up a group of counters to start on exec
was to set _all_ the counters to disabled and enable_on_exec. The way
this should be done is apparently to only set the group leader to
disabled and enable_on_exec.