perfgroup: A tool to measure performance of groups of processes
We sometimes need to measure the performance of groups of
processes. The perfgroup tool accomplishes this by providing roughly
the same features as rawperf, but instead of working on one process,
it starts a group of processes and measures performance for each of
them.

For the UART people: This is basically a hybrid between the old
benchrun tool and rawperf.