perfgroup: Support working directory specification
When running some benchmarks, it is useful to be able to specify a
working directory. This patch adds support for this using the
following syntax:

perfgroup -e hw:cpu_cycles -- @/path ./target

The working directory will be changed before the target application is
exec'ed, so the path to the target should be specified relative to the
specified working directory.