perfgroup: Catch fatal signals and try to do some cleanup
There are various reasons why perfgroup may receive a fatal
signal. One reason might be a failed assertion that causes a SIGABRT
to be raised. In this case, we need to make sure that all child
processes are terminated as well.

We accomplish this by installing a signal handler that sends a SIGTERM
to all child process groups and then raising the original signal. We
don't need to explicitly unmask and install the original handler for
the signal since sigaction, with the right flags, does this for us.