Remove unused code to send perf_event FDs
In an early version of the perf_common, it used to be possible to
start a child process, attach performance counters and send the
perf_event file descriptor to the parent using a socket. This method
fell into disuse in favor for a simpler method using sockets for
synchronization and setting up monitoring by attaching perf_events to
the target PID. Remove the now unused code for sending FDs between
processes.