Add a work around when pirating non-power-of-2 caches
When pirating caches with a non-power-of-2 size, the index function
sometimes uses bits outside of the range of the huge pages. This only
happens in some corner cases (at least on Westemere based systems with
a 12MB cache). Work around this issue by only using the first way-size
bytes of each huge page.